Move video scenes to different file to help test splitting
